requirements.txt
pyyaml>=6.0
requests>=2.28.0
pandas>=1.4.0
tqdm>=4.64.0
PySocks>=1.7.1 